dev: enhance snmp module

+++ component/rhilex_api_server/model/snmp_oids_point.go
@@ -23,5 +23,5 @@ type MSnmpOid struct {
 	Oid        string `gorm:"not null"` // .1.3.6.1.2.1.25.1.6.0
 	Tag        string `gorm:"not null"` // temp
 	Alias      string `gorm:"not null"` // 温度
-	Frequency  int    `gorm:"not null"` // 请求频率
+	Frequency  uint64 `gorm:"not null"` // 请求频率
 }
